analytics-service/requirements.txt
flask
flask-cors
camelot-py[cv]   # camelot with OpenCV backend
pandas
numpy